requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.1
django-extensions==3.1.3
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
graphviz==0.17
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7